+++ google-books-application/src/app/app.module.ts
@@ -3,7 +3,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ToastrModule } from 'ngx-toastr';
 import { BooksModule } from './modules/books/books.module';
 import { RouterModule } from '@angular/router';
 import { AuthInterceptor } from './core/auth/interceptors/auth.interceptor';
@@ -18,7 +17,6 @@ import { CoreModule } from './core/core.module';
     BrowserAnimationsModule,
     CoreModule,
     BooksModule,
-    ToastrModule.forRoot(),
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },