+++ google-books-application/src/app/core/header/components/header.component.ts
@@ -44,6 +44,7 @@ export class HeaderComponent implements OnInit {
             return;
           }
           this.userProfile = data;
+          console.log(this.userProfile, 'userProfile');
         },
       });
   }