+++ google-books-application/src/app/modules/books/components/book-card-list/components/card/book-card-list.component.ts
@@ -9,11 +9,11 @@ import {
 import { BookItem } from '../../../../models/book-item.interface';
 import { CARD_IMAGE_PLACEHOLDER } from '../../../../constants/card-image-placeholder.const';
 import { BooksService } from '../../../../services/books.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { LoaderService } from '../../../../../../shared/services/loader.service';
 import { finalize } from 'rxjs';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { FAVORITE_OPERATION_MESSAGE } from '../../../../constants/favorite-operation-message.const';
+import { NotificationService } from '../../../../../../shared/services/notification.service';
 
 @Component({
   selector: 'app-book-card-list',
@@ -22,30 +22,28 @@ import { FAVORITE_OPERATION_MESSAGE } from '../../../../constants/favorite-opera
 })
 export class BookCardListComponent {
   @Input({ required: true }) bookItems: BookItem[] = [];
-  @Output() favoriteStateChange: EventEmitter<void> = new EventEmitter<void>();
+  @Output() updateFavoriteShelf: EventEmitter<string> =
+    new EventEmitter<string>();
 
-  protected CARD_IMAGE_PLACEHOLDER = CARD_IMAGE_PLACEHOLDER;
+  protected readonly CARD_IMAGE_PLACEHOLDER = CARD_IMAGE_PLACEHOLDER;
   private readonly destroyRef: DestroyRef = inject(DestroyRef);
 
   constructor(
     private bookService: BooksService,
     private loaderService: LoaderService,
-    private matSnackbar: MatSnackBar
+    private notificationService: NotificationService
   ) {}
 
-  // TODO: Instead of dialog, open this book in official Google books page.
-  protected onOpenDetailDialog(): void {}
-
-  private openSnackbarWithMessage(message: string): void {
-    this.matSnackbar.open(message, undefined, {
-      duration: 3000,
-    });
+  protected onFavoriteIconClick(book: BookItem): void {
+    book.isFavorite
+      ? this.deleteBookFromFavoriteShelft(book.id)
+      : this.addBookToFavoriteShelf(book.id);
   }
 
-  protected onChangeFavoriteState(book: BookItem): void {
+  private addBookToFavoriteShelf(id: string): void {
     this.loaderService.start();
     this.bookService
-      .changeFavoriteStateById(!book.isFavorite, book.id)
+      .addBookToFavoriteById(id)
       .pipe(
         finalize(() => {
           this.loaderService.stop();
@@ -54,14 +52,33 @@ export class BookCardListComponent {
       )
       .subscribe({
         next: () => {
-          book.isFavorite = !book.isFavorite;
-          this.openSnackbarWithMessage(
-            book.isFavorite
-              ? FAVORITE_OPERATION_MESSAGE.ADDED
-              : FAVORITE_OPERATION_MESSAGE.REMOVED
+          this.notificationService.showSnackbar(
+            FAVORITE_OPERATION_MESSAGE.ADDED
           );
-          this.favoriteStateChange.emit();
+          this.updateFavoriteShelf.emit(id);
         },
       });
   }
+
+  private deleteBookFromFavoriteShelft(id: string): void {
+    this.loaderService.start();
+    this.bookService
+      .deleteBookFromFavoriteById(id)
+      .pipe(
+        finalize(() => {
+          this.loaderService.stop();
+        }),
+        takeUntilDestroyed(this.destroyRef)
+      )
+      .subscribe({
+        next: () => {
+          this.notificationService.showSnackbar(
+            FAVORITE_OPERATION_MESSAGE.REMOVED
+          );
+          this.updateFavoriteShelf.emit(id);
+        },
+      });
+  }
+
+  protected onOpenDetailDialog(): void {}
 }